scripts/comic/requirements.txt
# Comic QA Automation System Requirements
# Banking-Level Precision Dependencies

# Core image processing
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# Machine learning for advanced features
scikit-learn>=1.0.0

# Data processing and analysis
pandas>=1.3.0

# Optional: Deep learning for character recognition
# torch>=1.9.0
# torchvision>=0.10.0
# transformers>=4.0.0

# Development and testing
pytest>=6.0.0
pytest-cov>=2.0.0